A6: split SubmissionRepository into Reader + Writer behind a façade (#563 backlog)
Final and largest repo of the read/write split (instance-based, like
AppointmentRepository). SubmissionReader holds the 19 domain reads (lookups by
auth-code/token/email/cpf-rf/form, the user-link audit queries, export batching,
pagination, counts, the user-certificate-count subquery); SubmissionWriter holds
the 10 writes (insert/update overrides, status changes, bulk update/delete,
moveBetweenForms, deleteByFormId, QR-cache clear). Read-only helpers (hash,
build_export_where) and write-only helpers (count-cache invalidation, in-form
conflict check) go to their respective side; column_exists is duplicated because
both a read (hasEditInfo) and a write (updateWithEditTracking) need it.

The façade keeps the public static get_submissions_table(), the
get_table_name/get_cache_group/get_allowed_order_columns overrides, composes
reader+writer, delegates every domain method, and leaves the generic CRUD
inherited from AbstractRepository in place. All three bind the same global $wpdb,
so transactions stay coherent. Tech-debt note points to the future call-site
migration (#563 B3).

Behavior-preserving — Submission suite green (353 tests, incl. the reflection-
based repository + centralization tests); PHPStan L8 + WPCS clean.
